move deps_path
